Extracted post email serializer from mega service
