+++ core/server/services/mega/mega.js
@@ -1,24 +1,12 @@
-const juice = require('juice');
 const common = require('../../lib/common');
 const api = require('../../api');
 const membersService = require('../members');
 const bulkEmailService = require('../bulk-email');
 const models = require('../../models');
-const template = require('./template');
-const settingsCache = require('../../services/settings/cache');
-const urlUtils = require('../../lib/url-utils');
-
-const getSite = () => {
-    return Object.assign({}, settingsCache.getPublic(), {
-        url: urlUtils.urlFor('home', true)
-    });
-};
+const postEmailSerializer = require('./post-email-serializer');
 
 const sendEmail = async (post) => {
-    const emailTmpl = {
-        subject: post.email_subject || post.title,
-        html: juice(template({post, site: getSite()}))
-    };
+    const emailTmpl = postEmailSerializer.serialize(post);
 
     const {members} = await membersService.api.members.list();
     const emails = members.map(m => m.email);
